+++ src/service_worker.js
@@ -10,6 +10,7 @@ import {
   showNotificationTab,
   getAndStoreMetaData,
   overwriteSubmitMethod,
+  getDocumentDataFromDownloadModal,
 } from './utils/background.js';
 import { handleBackgroundFetch } from './utils/background_fetch.js';
 import { handleBackgroundNotification } from './utils/background_notifier.js';
@@ -35,6 +36,10 @@ chrome.runtime.onMessage.addListener((msg, sender, sendResponse) => {
   }
   if (msg.message === 'getMetaData') {
     getAndStoreMetaData(msg, sender, sendResponse);
+    return true;
+  }
+  if (msg.message === 'getDocumentDataFromDownloadModal') {
+    getDocumentDataFromDownloadModal(msg, sender, sendResponse);
     return true;
   }
   if (msg.message === 'overwriteSubmit') {